backend/requirements.txt
langchain_ollama
langchain_core
langchain_community
langchain
langchain_groq
langchain_huggingface
langchain_text_splitters
python-dotenv
streamlit
pypdf
faiss-cpu
streamlit
fastapi
uvicorn
langchain
langchain-community
langchain-core
langchain-groq
langchain-huggingface
sentence-transformers
faiss-cpu
pypdf
python-dotenv
python-multipart